Update Axios version to 0.27.2 and add useAxios0_27_2 flag
This commit updates the Axios version to 0.27.2 in multiple files and adds a new flag called useAxios0_27_2. This flag can be used to control whether to use Axios 0.27.2 or the latest version. The useAxios0_27_2 flag is added to the generator configuration, TypeScript generation function, JavaGenerateApiRequestBody, KonfigYaml, and TypeScriptAxiosClientCodegen files. This change allows for publishing the TypeScript SDK without making breaking changes.